Migrate PredefinedStyles to common lib
