+++ lib/PredefinedStyleLib/PredefinedStyleLib.h
@@ -1,0 +1,11 @@
+#ifndef PREDEFINEDSTYLELIB_H
+#define PREDEFINEDSTYLELIB_H
+
+#include <Arduino.h>
+#include <PatternData.h>
+#include <vector>
+#include "PredefinedStyle.h"
+#include "PredefinedStyles.h"
+#include "PredefinedStyleList.h"
+
+#endif
