+++ lib/PredefinedStyleLib/PredefinedStyles.h
@@ -1,0 +1,16 @@
+#ifndef PREDEFINEDSTYLES_H
+#define PREDEFINEDSTYLES_H
+
+enum class PredefinedStyles {
+    Pink_Solid,
+    RedPink_Right,
+    BluePink_Right,
+    Rainbow_Right,
+    RedPink_Random,
+    BluePink_Random,
+    Rainbow_Random,
+    RedPink_CenterOut,
+    BluePink_CenterOut
+};
+
+#endif
